+++ spec/configuration/Basic.spec.php
@@ -43,7 +43,7 @@ describe('Basic', function() {
             expect($this->configration->getService())->toBeAnInstanceOf('\coverallskit\entity\ServiceInterface');
         });
         it('set the repository', function() {
-            expect($this->configration->getRepository())->toBeAnInstanceOf('\coverallskit\entity\RepositoryInterface');
+            expect($this->configration->getRepository())->toBeAnInstanceOf('\coverallskit\entity\RepositoryEntity');
         });
     });
     describe('applyTo', function() {
@@ -58,7 +58,7 @@ describe('Basic', function() {
             expect($this->builder->getService())->toBeAnInstanceOf('coverallskit\entity\ServiceInterface');
         });
         it('apply repository config', function() {
-            expect($this->builder->getRepository())->toBeAnInstanceOf('coverallskit\entity\RepositoryInterface');
+            expect($this->builder->getRepository())->toBeAnInstanceOf('coverallskit\entity\RepositoryEntity');
         });
     });
 });