+++ spec/entity/SourceFile.spec.php
@@ -12,7 +12,7 @@
 namespace coverallskit\spec;
 
 use coverallskit\entity\SourceFile;
-use coverallskit\entity\Coverage;
+use coverallskit\entity\CoverageResult;
 
 describe('SourceFile', function() {
     beforeEach(function() {
@@ -70,7 +70,7 @@ describe('SourceFile', function() {
     });
     describe('addCoverage', function() {
         beforeEach(function() {
-            $this->coverage = Coverage::unused(1);
+            $this->coverage = CoverageResult::unused(1);
             $this->sourceFile = new SourceFile($this->path);
             $this->sourceFile->addCoverage($this->coverage);
             $this->retrieveCoverage = $this->sourceFile->getCoverage(1);
@@ -82,14 +82,14 @@ describe('SourceFile', function() {
         context('when line out of range', function() {
             it('should add coverage', function() {
                 expect(function() {
-                    $coverage = Coverage::unused(999);
+                    $coverage = CoverageResult::unused(999);
                     $this->sourceFile->addCoverage($coverage);
                 })->toThrow('coverallskit\exception\LineOutOfRangeException');
             });
         });
         context('when the blank line of the last', function() {
             it('should add coverage', function() {
-                $coverage = Coverage::unused(4);
+                $coverage = CoverageResult::unused(4);
                 $this->sourceFile->addCoverage($coverage);
 
                 expect($this->sourceFile->getCoverage(21))->toBeNull();
@@ -99,7 +99,7 @@ describe('SourceFile', function() {
 
     describe('removeCoverage', function() {
         beforeEach(function() {
-            $this->coverage = Coverage::unused(3);
+            $this->coverage = CoverageResult::unused(3);
             $this->sourceFile->addCoverage($this->coverage);
         });
         it('should add coverage', function() {
@@ -111,8 +111,8 @@ describe('SourceFile', function() {
     describe('getExecutedLineCount', function() {
         beforeEach(function() {
             $this->sourceFile = new SourceFile($this->path);
-            $this->sourceFile->addCoverage(Coverage::executed(12));
-            $this->sourceFile->addCoverage(Coverage::unused(17));
+            $this->sourceFile->addCoverage(CoverageResult::executed(12));
+            $this->sourceFile->addCoverage(CoverageResult::unused(17));
         });
         it('return executed line count', function() {
             expect($this->sourceFile->getExecutedLineCount())->toEqual(1);
@@ -122,8 +122,8 @@ describe('SourceFile', function() {
     describe('getUnusedLineCount', function() {
         beforeEach(function() {
             $this->sourceFile = new SourceFile($this->path);
-            $this->sourceFile->addCoverage(Coverage::executed(12));
-            $this->sourceFile->addCoverage(Coverage::unused(17));
+            $this->sourceFile->addCoverage(CoverageResult::executed(12));
+            $this->sourceFile->addCoverage(CoverageResult::unused(17));
         });
         it('return unused line count', function() {
             expect($this->sourceFile->getUnusedLineCount())->toEqual(1);