+++ src/configuration/Basic.php
@@ -13,7 +13,7 @@ namespace coverallskit\configuration;
 
 use coverallskit\ReportBuilderInterface;
 use coverallskit\entity\Service;
-use coverallskit\entity\Repository;
+use coverallskit\entity\GitRepository;
 use coverallskit\Environment;
 use coverallskit\environment\AdaptorResolver;
 use Zend\Config\Config;
@@ -74,7 +74,7 @@ class Basic extends AbstractConfiguration
      }
 
     /**
-     * @return \coverallskit\entity\Repository
+     * @return \coverallskit\entity\RepositoryEntity
      */
     public function getRepository()
     {
@@ -99,12 +99,12 @@ class Basic extends AbstractConfiguration
 
     /**
      * @param string $path
-     * @return \coverallskit\entity\Repository
+     * @return \coverallskit\entity\RepositoryEntity
      */
     private function repositoryFromPath($path)
     {
         $directory = $this->resolvePath($path);
-        $repository = new Repository($directory);
+        $repository = new GitRepository($directory);
 
         return $repository;
     }